Handle custom and echo mission in same files
