+++ backend/api.test/TestWebApplicationFactory.cs
@@ -36,7 +36,6 @@ namespace Api.Test
                     services.AddScoped<IMapService, MockMapService>();
                     services.AddScoped<IBlobService, MockBlobService>();
                     services.AddScoped<IStidService, MockStidService>();
-                    services.AddScoped<ICustomMissionService, MockCustomMissionService>();
                     services.AddAuthorizationBuilder().AddFallbackPolicy(
                         TestAuthHandler.AuthenticationScheme, policy => policy.RequireAuthenticatedUser()
                     );
